requirements.txt
Django>=4.0
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
